xcb: Process all pending events before repainting
We used to process them one at a time and, when we came across a
ATOM_NOTICE client message, we would repaint on the spot.  The problem was
that every CONFIGURE cased an additional ATOM_NOTICE client message to get
inserted into the stream and we would get one frame behind.  Throw in
enough configure events and it starts to lag pretty bad.