Use vkGetInstanceProcAddr to find wayland specific functions
According to the documentation of vkGetDeviceProcAddr [1], it may be
used only for functions that take a VkDevice or one of its child object
types as its first argument. The returned function calls directly into
the ICD, bypassing the object unwrapping done in the Vulkan loader
dispatch functions.
When linked against the Vulkan loader, vkEnumeratePhysicalDevices returns
wrapped vkPhysicalDevices, so this used to cause a segmentation fault
in vkGetPhysicalDeviceWaylandPresentationSupportKHR.

[1] https://www.khronos.org/registry/vulkan/specs/1.0/man/html/vkGetDeviceProcAddr.html