main: make initialization work with LunarG's loader
It seems we'll crash into the loader's trampoline if we don't query symbols
using the instance rather than the device.

Also wayland or xcb support symbols requires extensions related to surfaces
that imply VK_KHR_surface. The loader doesn't seem to understand the
dependency, therefore we might want to put VK_KHR_surface in the list of
extension we need.