Make init funcs return -1 on failure
That is, init_headless(), init_kms(), init_wayland(), and init_xcb().
This allows us to explicitly check if initialization failed instead of
checking for failure by inspecting global variables.

This cleanup makes the patch a LOT easier to write. It adds a cmdline
option to select the display mode (wayland, kms, etc).