Explicitly track the display mode
This adds a global variable 'display_mode'.
Its initial value is DISPLAY_MODE_AUTO, which preserves the
default behavior of trying Wayland, XCB, KMS, and headless, in that
order.

Once initialization of some mode succeeds, update the global display_mode
and use that to choose which mainloop to run. Pre-patch, we chose which
mainloop to run by probing several different global variables.

This makes the next patch a LOT easier to write, which adds a cmdline
option to select a specific display mode.