Pick a usable memory type for the vertex buffer
Previously it was hardcoded to use memory type 0. Since 39adea9 in the
Mesa Intel Vulkan driver this triggers an assert because the first
memory type doesn’t support VK_BUFFER_USAGE_VERTEX_BUFFER_BIT. To fix
that it now checks the usable memory types with
vkGetBufferMemoryRequirements and picks the first one.