wait for and reset image-fence before rendering
Since we have a fence per image, this shouldn't stall unless we're
rendering faster than the presentation rate anyway. So this
shouldn't lead to stalling.

This gets rid of a validation-error on the form:

Fence 0x16 is already in use by another submission.

Attempt to reset command pool with command buffer which is in
use. The spec valid usage text states 'All VkCommandBuffer
objects allocated from commandPool must not be in the pending
state'

https://www.khronos.org/registry/vulkan/specs/1.0/html/vkspec.html#VUID-vkResetCommandPool-commandPool-00040