respect surface_caps.{min,max}ImageCount
This gets rid of a validation error on this form on my machine:

vkCreateSwapChainKHR() called with minImageCount = 2, which is
outside the bounds returned by vkGetPhysicalDeviceSurfaceCapabilitiesKHR()
(i.e. minImageCount = 4, maxImageCount = 0). The spec valid usage
text states 'minImageCount must be greater than or equal to the
value returned in the minImageCount member of the VkSurfaceCapabilitiesKHR
structure returned by vkGetPhysicalDeviceSurfaceCapabilitiesKHR for
the surface

https://www.khronos.org/registry/vulkan/specs/1.0-extensions/html/vkspec.html#VUID-VkSwapchainCreateInfoKHR-minImageCount-01271